Refactor getForecasts method into a Darksky service
